Build 'Create' new item feature
Appended showStateAllTemplate and showStateItemCreateTemplate handlebar
templates to state-view container in getItemsSuccess. This allowed us to
create a 'submit' handler for the submission event of a newly created
item. Now when a user clicks submit, the onCreateItem function is run,
sending a POST request to the API with the 'region' (alias for
the states) stored from when user initially clicked on the state
(store.state = region).

Note: In the onCreateItem, we created a 'newData' object that contains
the user-inputted content (description, due_date, status, title,
category, and location) and the stored state. In the API call, we pass
in this 'newData' object because our API requires all attributes,
including state.

Worked with Nikki Riser @nriser