Add a 'back to map' button
Add a 'back to map' button in the 'state' view. When a user clicks on
the button, the user can now return to the 'map' view.

Bug fix: The state that was clicked initially on the map page
retains its on-click color after returning to the 'map' view from the
'state' view. Changed the default styling in the usMap function to fix
this bug.