But fix for apparently non-functional buttons
Fixed two buttons that didn't appear to do anything in certain situations:

1. When user clicks on a state with no items, the add item button doesn't
seem to do anything if clicked because the create item form is already
displayed by default.  The button would actually refresh that form, but
that was not obvious, so the button has been removed whenever the create
item form is displayed.

2. If the user has no items in a state and clicks "cancel" on the create
item field, again nothing would happen.  Removed the "cancel" button in
that situation.