Fixed bug relating to create/cancel buttons in state view
We changed the behavior of the show item form on the right side of the
state view page.  As before, if there are incomplete items, it will show
the incomplete item with the earliest due date.  Now, after it checks for
incomplete items, it will also check for completed items.  If there are
completed items, it will show the completed item with the most recent due
date.  If there are no items (completed or incomplete) then it displays
the create item form.  This fixed the following bugs:

1. For a state: when you only have completed items. when you click
details you cant create a new item. There's no + button.
2. If you have only completed items and you click show details, then
click edit, then cancel, then create an item, from there clicking the
cancel button would simply remove the cancel button itself and not
appear to do anything else.